Improved removal
Please note that checking if something went off the screen is NOT the
only way to determine is something is dead, nor is it necessarily the
best. This is a simple example. If you want to have multiple conditions
that kill a particle, let me know, and I can explain how.